# api/openapi3.yaml
components:
    parameters:
        AuthorizationTokenHPN:
            description: Correlation id
            in: header
            name: Authorization
            schema:
                format: uuid
                type: string
        CorrelationIDHPN:
            description: Correlation id
            in: header
            name: X-Request-ID
            schema:
                format: uuid
                type: string
        IDPPN:
            in: path
            name: id
            required: true
            schema:
                format: uuid
                type: string
    requestBodies:
        CreateRecordRequest:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Record'
            description: Request used for creating a record.
            required: true
        UpdateRecordRequest:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Record'
            description: Request used for updating a record.
            required: true
    responses:
        ErrorResponse:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Response when errors happen.
        ListRecordsResponse:
            content:
                application/json:
                    schema:
                        items:
                            $ref: '#/components/schemas/Record'
                        type: array
            description: Response returns back all records.
        RecordResponse:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Record'
            description: Response returns back successfully found or created record.
    schemas:
        Error:
            properties:
                message:
                    type: string
            type: object
        Record:
            properties:
                created_at:
                    format: date-time
                    type: string
                created_by:
                    type: string
                description:
                    type: string
                id: {}
                login:
                    type: string
                name:
                    type: string
                password:
                    type: string
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    type: string
                url:
                    type: string
            type: object
        User:
            properties:
                created_at:
                    format: date-time
                    type: string
                email:
                    type: string
                id: {}
                login:
                    type: string
                name:
                    type: string
                password:
                    type: string
                phone:
                    type: string
                updated_at:
                    format: date-time
                    type: string
            type: object
    securitySchemes:
        AuthorizationToken:
            bearerFormat: JWT
            scheme: bearer
            type: http
info:
    contact:
        url: https://github.com/okutsen/PasswordManager
    title: Password Manager
    version: 0.0.0
openapi: 3.0.0
paths:
    /records:
        get:
            operationId: ListRecords
            responses:
                "200":
                    $ref: '#/components/responses/ListRecordsResponse'
                "500":
                    $ref: '#/components/responses/ErrorResponse'
        post:
            operationId: CreateRecord
            requestBody:
                $ref: '#/components/requestBodies/CreateRecordRequest'
            responses:
                "201":
                    $ref: '#/components/responses/RecordResponse'
                "400":
                    $ref: '#/components/responses/ErrorResponse'
                "500":
                    $ref: '#/components/responses/ErrorResponse'
    /records/{id}:
        delete:
            operationId: DeleteRecord
            parameters:
                - $ref: '#/components/parameters/IDPPN'
            responses:
                "200":
                    description: Record deleted
                "400":
                    $ref: '#/components/responses/ErrorResponse'
                "500":
                    $ref: '#/components/responses/ErrorResponse'
        get:
            operationId: GetRecord
            parameters:
                - $ref: '#/components/parameters/IDPPN'
            responses:
                "200":
                    $ref: '#/components/responses/RecordResponse'
                "400":
                    $ref: '#/components/responses/ErrorResponse'
                "500":
                    $ref: '#/components/responses/ErrorResponse'
        put:
            operationId: UpdateRecord
            parameters:
                - $ref: '#/components/parameters/IDPPN'
            requestBody:
                $ref: '#/components/requestBodies/UpdateRecordRequest'
            responses:
                "202":
                    $ref: '#/components/responses/RecordResponse'
                "400":
                    $ref: '#/components/responses/ErrorResponse'
                "500":
                    $ref: '#/components/responses/ErrorResponse'
servers:
    - description: Local development
      url: http://127.0.0.1:10000